pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "truenas_pylibvirt"
version = "0.0.1"
description = "TrueNAS libvirt wrapper"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "libvirt-python>=9.0.0",
    "pyudev>=0.24.0",
    # truenas-pynetif is not on PyPI, must be installed separately
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["truenas_pylibvirt*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
truenas_pylibvirt = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
strict = true

[[tool.mypy.overrides]]
module = ["libvirt.*", "pyudev.*"]
ignore_missing_imports = true
